This implementation on hold for a compile time implementation
