+++ libs/CMakeLists.txt
@@ -5,6 +5,7 @@ add_subdirectory(confman)
 add_subdirectory(ipc)
 add_subdirectory(rpc)
 add_subdirectory(stx)
+add_subdirectory(rpc)
 
 add_library(tfc INTERFACE)
 add_library(tfc::tfc ALIAS tfc)
@@ -14,4 +15,5 @@ target_link_libraries(tfc
     tfc::confman,
     tfc::base,
     tfc::ipc,
+    tfc::rpc,
 )
